:recycle: Refactor logging
